requirements.txt
fastapi~=0.115.0
fastapi-cli~=0.0.5
icecream~=2.1.3
sqlmodel~=0.0.22
numpy~=1.26.4
pydantic~=2.9.2
SQLAlchemy~=2.0.35
httpx~=0.27.2
starlette~=0.40.0
asyncpg~=0.29.0
python-dotenv~=1.0.1
minio~=7.2.9
python-multipart~=0.0.12
greenlet~=3.1.1
